refactor(desktop): inline diff stats reducer back into list item
The useLocalDiffStats hook was only called in one place; the other three
imports were just for the LocalDiffStats type on a prop. Inline the
useMemo back into WorkspaceListItem and use a structural type on the
prop instead — drops the hook file and its index export.
