+++ apps/desktop/src/renderer/screens/main/hooks/index.ts
@@ -1,6 +1,5 @@
 export { useBranchSyncInvalidation } from "./useBranchSyncInvalidation";
 export { useCreateOrOpenPR } from "./useCreateOrOpenPR";
 export { useGitChangesStatus } from "./useGitChangesStatus";
-export { type LocalDiffStats, useLocalDiffStats } from "./useLocalDiffStats";
 export { usePRStatus } from "./usePRStatus";
 export { useWorkspaceRename } from "./useWorkspaceRename";
