--- apps/desktop/src/renderer/screens/main/hooks/useLocalDiffStats/index.ts
@@ -1,1 +1,0 @@
-export { type LocalDiffStats, useLocalDiffStats } from "./useLocalDiffStats";
